+++ config/config_test.go
@@ -18,26 +18,16 @@ limitations under the License.
 package config_test
 
 import (
-	"strconv"
+	"errors"
+	"strings"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/jthomperoo/custom-pod-autoscaler/config"
-	"github.com/jthomperoo/custom-pod-autoscaler/cpatest"
-	"gopkg.in/yaml.v2"
+	v1 "k8s.io/api/autoscaling/v1"
 )
 
 const (
-	evaluateEnvName        = "evaluate"
-	metricEnvName          = "metric"
-	intervalEnvName        = "interval"
-	hostEnvName            = "host"
-	portEnvName            = "port"
-	metricTimeoutEnvName   = "metricTimeout"
-	evaluateTimeoutEnvName = "evaluateTimeout"
-	namespaceEnvName       = "namespace"
-	scaleTargetRefEnvName  = "scaleTargetRef"
-
 	defaultEvaluate        = ">&2 echo 'ERROR: No evaluate command set' && exit 1"
 	defaultMetric          = ">&2 echo 'ERROR: No metric command set' && exit 1"
 	defaultInterval        = 15000
@@ -46,121 +36,141 @@ const (
 	defaultMetricTimeout   = 5000
 	defaultEvaluateTimeout = 5000
 	defaultNamespace       = "default"
-
-	invalidYAML                   = "- in: -: valid - yaml"
-	testEvaluate                  = "test evaluate"
-	testMetric                    = "test metric"
-	testInterval                  = 1234
-	testHost                      = "1.2.3.4"
-	testPort                      = 1234
-	testMetricTimeout             = 4321
-	testEvaluateTimeout           = 8765
-	testNamespace                 = "test namespace"
-	testScaleTargetRefKind        = "test kind"
-	testScaleTargetRefName        = "test name"
-	testScaleTargetRefAPIVersion  = "test api version"
-	testScaleTargetRefJSON        = "{\"kind\":\"test kind\", \"name\":\"test name\", \"apiVersion\":\"test api version\" }"
-	testScaleTargetRefInvalidJSON = "abc invali:d json"
 )
 
-func TestLoadConfig_InvalidYAML(t *testing.T) {
-	_, err := config.LoadConfig([]byte(invalidYAML), nil)
-	if err == nil {
-		t.Errorf("Expected error due to invalid YAML provided")
-	}
-}
-
-func TestLoadConfig_InvalidIntEnv(t *testing.T) {
-	_, err := config.LoadConfig(nil, map[string]string{
-		intervalEnvName: "invalid",
+func TestLoadConfig(t *testing.T) {
+	equateErrorMessage := cmp.Comparer(func(x, y error) bool {
+		if x == nil || y == nil {
+			return x == nil && y == nil
+		}
+		return x.Error() == y.Error()
 	})
 
-	if err == nil {
-		t.Errorf("Expected error due to invalid integer environment variable provided")
+	var tests = []struct {
+		description string
+		yamlInput   []byte
+		envVars     map[string]string
+		expectedErr error
+		expected    *config.Config
+	}{
+		{
+			"Invalid YAML",
+			[]byte("invalid"),
+			nil,
+			errors.New("yaml: unmarshal errors:\n  line 1: cannot unmarshal !!str `invalid` into config.Config"),
+			nil,
+		},
+		{
+			"Invalid int YAML config",
+			[]byte("interval: \"invalid\""),
+			nil,
+			errors.New("yaml: unmarshal errors:\n  line 1: cannot unmarshal !!str `invalid` into int"),
+			nil,
+		},
+		{
+			"Invalid int env config",
+			nil,
+			map[string]string{
+				"interval": "invalid",
+			},
+			errors.New("strconv.ParseInt: parsing \"invalid\": invalid syntax"),
+			nil,
+		},
+		{
+			"Invalid scale target ref",
+			nil,
+			map[string]string{
+				"scaleTargetRef": "invalid JSON",
+			},
+			errors.New("invalid character 'i' looking for beginning of value"),
+			nil,
+		},
+		{
+			"No YAML no env return default",
+			nil,
+			nil,
+			nil,
+			&config.Config{
+				Evaluate:        defaultEvaluate,
+				Metric:          defaultMetric,
+				Interval:        defaultInterval,
+				Host:            defaultHost,
+				Port:            defaultPort,
+				MetricTimeout:   defaultMetricTimeout,
+				EvaluateTimeout: defaultEvaluateTimeout,
+				Namespace:       defaultNamespace,
+				ScaleTargetRef:  nil,
+			},
+		},
+		{
+			"No YAML override with env",
+			nil,
+			map[string]string{
+				"evaluate":        "test env evaluate",
+				"metric":          "test env metric",
+				"interval":        "35",
+				"host":            "test env host",
+				"port":            "1234",
+				"metricTimeout":   "13",
+				"evaluateTimeout": "14",
+				"namespace":       "test env namespace",
+				"scaleTargetRef":  `{ "name": "test target name", "kind": "test target kind", "apiVersion": "test target api version"}`,
+			},
+			nil,
+			&config.Config{
+				Evaluate:        "test env evaluate",
+				Metric:          "test env metric",
+				Interval:        35,
+				Host:            "test env host",
+				Port:            1234,
+				MetricTimeout:   13,
+				EvaluateTimeout: 14,
+				Namespace:       "test env namespace",
+				ScaleTargetRef: &v1.CrossVersionObjectReference{
+					Name:       "test target name",
+					Kind:       "test target kind",
+					APIVersion: "test target api version",
+				},
+			},
+		},
+		{
+			"No env override with YAML",
+			[]byte(strings.Replace(`
+				evaluate: "test yaml evaluate"
+				metric: "test yaml metric"
+				interval: 30
+				host: "test yaml host"
+				port: 7890
+				metricTimeout: 10
+				evaluateTimeout: 11
+				namespace: "test yaml namespace"
+			`, "\t", "", -1)),
+			nil,
+			nil,
+			&config.Config{
+				Evaluate:        "test yaml evaluate",
+				Metric:          "test yaml metric",
+				Interval:        30,
+				Host:            "test yaml host",
+				Port:            7890,
+				MetricTimeout:   10,
+				EvaluateTimeout: 11,
+				Namespace:       "test yaml namespace",
+			},
+		},
 	}
-}
+	for _, test := range tests {
+		t.Run(test.description, func(t *testing.T) {
+			config, err := config.LoadConfig(test.yamlInput, test.envVars)
+			if !cmp.Equal(&err, &test.expectedErr, equateErrorMessage) {
+				t.Errorf("Error mismatch (-want +got):\n%s", cmp.Diff(test.expectedErr, err, equateErrorMessage))
+				return
+			}
 
-func TestLoadConfig_NoYAML(t *testing.T) {
-	testConfig := cpatest.GetTestConfig()
-
-	testEnvVars := getTestEnvVars()
-
-	loadedConfig, err := config.LoadConfig(nil, testEnvVars)
-	if err != nil {
-		t.Error(err)
-	}
-
-	if !cmp.Equal(loadedConfig, testConfig) {
-		t.Errorf("Config mismatch (-want +got):\n%s", cmp.Diff(testConfig, loadedConfig))
-	}
-}
-
-func TestLoadConfig_NoEnv(t *testing.T) {
-	testConfig := cpatest.GetTestConfig()
-
-	yamlConfig, err := yaml.Marshal(testConfig)
-	if err != nil {
-		t.Error(err)
-	}
-
-	loadedConfig, err := config.LoadConfig(yamlConfig, nil)
-	if err != nil {
-		t.Error(err)
-	}
-
-	if !cmp.Equal(loadedConfig, testConfig) {
-		t.Errorf("Config mismatch (-want +got):\n%s", cmp.Diff(testConfig, loadedConfig))
-	}
-}
-
-func TestLoadConfig_NoYAMLNoEnv(t *testing.T) {
-	testConfig := getDefaultConfig()
-
-	loadedConfig, err := config.LoadConfig(nil, nil)
-	if err != nil {
-		t.Error(err)
-	}
-
-	if !cmp.Equal(loadedConfig, testConfig) {
-		t.Errorf("Config mismatch (-want +got):\n%s", cmp.Diff(testConfig, loadedConfig))
-	}
-}
-
-func TestLoadCOnfig_InvalidScaleTargetRefJSON(t *testing.T) {
-	testEnvVars := getTestEnvVars()
-
-	testEnvVars[scaleTargetRefEnvName] = testScaleTargetRefInvalidJSON
-
-	_, err := config.LoadConfig(nil, testEnvVars)
-	if err == nil {
-		t.Errorf("Expected error due to invalid scaleTargetRef environment variable provided")
-	}
-}
-
-func getDefaultConfig() *config.Config {
-	return &config.Config{
-		Evaluate:        defaultEvaluate,
-		Metric:          defaultMetric,
-		Interval:        defaultInterval,
-		Host:            defaultHost,
-		Port:            defaultPort,
-		MetricTimeout:   defaultMetricTimeout,
-		EvaluateTimeout: defaultEvaluateTimeout,
-		Namespace:       defaultNamespace,
-		ScaleTargetRef:  nil,
-	}
-}
-
-func getTestEnvVars() map[string]string {
-	return map[string]string{
-		evaluateEnvName:        testEvaluate,
-		metricEnvName:          testMetric,
-		intervalEnvName:        strconv.FormatInt(testInterval, 10),
-		hostEnvName:            testHost,
-		portEnvName:            strconv.FormatInt(testPort, 10),
-		metricTimeoutEnvName:   strconv.FormatInt(testMetricTimeout, 10),
-		evaluateTimeoutEnvName: strconv.FormatInt(testEvaluateTimeout, 10),
-		namespaceEnvName:       testNamespace,
-		scaleTargetRefEnvName:  testScaleTargetRefJSON,
+			if !cmp.Equal(config, test.expected) {
+				t.Errorf("Config mismatch (-want +got):\n%s", cmp.Diff(test.expected, config))
+				return
+			}
+		})
 	}
 }